% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plUpliftEval.R
\name{pl_plot}
\alias{pl_plot}
\title{A port of pylift's plot function (https://github.com/wayfair/pylift) as of commit:
https://github.com/wayfair/pylift/tree/bb69692388b1fe085001c3ba7edf6dd81d888353
pylift: Plots the different kinds of percentage-targeted curves.}
\usage{
pl_plot(plue, plot_type = "cgains", n_bins = 20,
  show_theoretical_max = FALSE, show_practical_max = FALSE,
  show_random_selection = TRUE, show_no_dogs = FALSE, ...)
}
\arguments{
\item{plue}{the result of a call to the plUpliftEval constructor}

\item{plot_type}{string, optional Either 'qini', 'aqini', 'uplift', 'cuplift', or 'balance'.
'aqini' refers to an adjusted qini plot, 'cuplift' gives a
cumulative uplift plot. 'balance' gives the test-control balance
for each of the bins. All others are self-explanatory.}

\item{show_practical_max}{: boolean, optional
Toggle theoretical maximal qini curve, if not overfitting to
treatment/control. Only works for Qini-style curves.}

\item{show_no_dogs}{: boolean, optional
Toggle theoretical maximal qini curve, if you believe there are no
sleeping dogs. Only works for Qini-style curves.}

\item{...}{additional arguments}

\item{show_theoretical_max:}{boolean, optional
Toggle theoretical maximal qini curve, if overfitting to
treatment/control. Only works for Qini-style curves.}

\item{show_random_selection:}{boolean, optional
Toggle straight line indicating a random ordering. Only works for
Qini-style curves.}
}
\description{
A port of pylift's plot function (https://github.com/wayfair/pylift) as of commit:
https://github.com/wayfair/pylift/tree/bb69692388b1fe085001c3ba7edf6dd81d888353
pylift: Plots the different kinds of percentage-targeted curves.
}
