% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plUpliftEval.R
\name{plUpliftEval}
\alias{plUpliftEval}
\title{A helper for the new_PlUpliftEval function that validates the treatment,
outcome, prediction, p, and n_bins arguments.}
\usage{
plUpliftEval(treatment, outcome, prediction, p = "infer", n_bins = 20)
}
\arguments{
\item{treatment}{numeric vector of treatment identifiers}

\item{outcome}{numeric vector of outcomes}

\item{prediction}{numeric vector of uplift predictions}

\item{p}{optional "infer", numeric, numeric vector representing treatment
propensities}

\item{n_bins}{integer number of bins on x-axis; default 20}
}
\description{
A helper for the new_PlUpliftEval function that validates the treatment,
outcome, prediction, p, and n_bins arguments.
}
